Discordpy bump

+++ bot.py
@@ -1,4 +1,4 @@
-__version__ = "3.4.1"
+__version__ = "3.5.0"
 
 
 import asyncio
@@ -18,8 +18,6 @@ import isodate
 
 from aiohttp import ClientSession
 from emoji import UNICODE_EMOJI
-from motor.motor_asyncio import AsyncIOMotorClient
-from pymongo.errors import ConfigurationError
 
 from pkg_resources import parse_version
 
@@ -32,7 +30,7 @@ except ImportError:
     pass
 
 from core import checks
-from core.clients import ApiClient, PluginDatabaseClient
+from core.clients import ApiClient, PluginDatabaseClient, MongoDBClient
 from core.config import ConfigManager
 from core.utils import human_join, normalize_alias
 from core.models import PermissionLevel, SafeFormatter, getLogger, configure_logging
@@ -45,6 +43,12 @@ logger = getLogger(__name__)
 temp_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "temp")
 if not os.path.exists(temp_dir):
     os.mkdir(temp_dir)
+
+if sys.platform == "win32":
+    try:
+        asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
+    except AttributeError:
+        logger.error("Failed to use WindowsProactorEventLoopPolicy.", exc_info=True)
 
 
 class ModmailBot(commands.Bot):
@@ -66,22 +70,7 @@ class ModmailBot(commands.Bot):
         self.log_file_name = os.path.join(temp_dir, f"{self.token.split('.')[0]}.log")
         self._configure_logging()
 
-        mongo_uri = self.config["mongo_uri"]
-        if mongo_uri is None:
-            logger.critical("A Mongo URI is necessary for the bot to function.")
-            raise RuntimeError
-
-        try:
-            self.db = AsyncIOMotorClient(mongo_uri).modmail_bot
-        except ConfigurationError as e:
-            logger.critical(
-                "Your MONGO_URI might be copied wrong, try re-copying from the source again. "
-                "Otherwise noted in the following message:"
-            )
-            logger.critical(e)
-            sys.exit(0)
-
-        self.plugin_db = PluginDatabaseClient(self)
+        self.plugin_db = PluginDatabaseClient(self)  # Deprecated
         self.startup()
 
     @property
@@ -152,8 +141,17 @@ class ModmailBot(commands.Bot):
     @property
     def api(self) -> ApiClient:
         if self._api is None:
-            self._api = ApiClient(self)
+            if self.config["database_type"].lower() == "mongodb":
+                self._api = MongoDBClient(self)
+            else:
+                logger.critical("Invalid database type.")
+                raise RuntimeError
         return self._api
+
+    @property
+    def db(self):
+        # deprecated
+        return self.api.db
 
     async def get_prefix(self, message=None):
         return [self.prefix, f"<@{self.user.id}> ", f"<@!{self.user.id}> "]
@@ -180,7 +178,7 @@ class ModmailBot(commands.Bot):
                 logger.error(" - Shutting down bot - ")
 
     @property
-    def owner_ids(self):
+    def bot_owner_ids(self):
         owner_ids = self.config["owners"]
         if owner_ids is not None:
             owner_ids = set(map(int, str(owner_ids).split(",")))
@@ -192,7 +190,7 @@ class ModmailBot(commands.Bot):
         return owner_ids
 
     async def is_owner(self, user: discord.User) -> bool:
-        if user.id in self.owner_ids:
+        if user.id in self.bot_owner_ids:
             return True
         return await super().is_owner(user)
 
@@ -370,36 +368,15 @@ class ModmailBot(commands.Bot):
 
     async def on_connect(self):
         try:
-            await self.validate_database_connection()
+            await self.api.validate_database_connection()
         except Exception:
             logger.debug("Logging out due to failed database connection.")
             return await self.logout()
 
         logger.debug("Connected to gateway.")
         await self.config.refresh()
-        await self.setup_indexes()
+        await self.api.setup_indexes()
         self._connected.set()
-
-    async def setup_indexes(self):
-        """Setup text indexes so we can use the $search operator"""
-        coll = self.db.logs
-        index_name = "messages.content_text_messages.author.name_text_key_text"
-
-        index_info = await coll.index_information()
-
-        # Backwards compatibility
-        old_index = "messages.content_text_messages.author.name_text"
-        if old_index in index_info:
-            logger.info("Dropping old index: %s", old_index)
-            await coll.drop_index(old_index)
-
-        if index_name not in index_info:
-            logger.info('Creating "text" index for logs collection.')
-            logger.info("Name: %s", index_name)
-            await coll.create_index(
-                [("messages.content", "text"), ("messages.author.name", "text"), ("key", "text")]
-            )
-        logger.debug("Successfully configured and verified database indexes.")
 
     async def on_ready(self):
         """Bot startup, sets uptime."""
@@ -416,7 +393,8 @@ class ModmailBot(commands.Bot):
         logger.info("Logged in as: %s", self.user)
         logger.info("Bot ID: %s", self.user.id)
         owners = ", ".join(
-            getattr(self.get_user(owner_id), "name", str(owner_id)) for owner_id in self.owner_ids
+            getattr(self.get_user(owner_id), "name", str(owner_id))
+            for owner_id in self.bot_owner_ids
         )
         logger.info("Owners: %s", owners)
         logger.info("Prefix: %s", self.prefix)
@@ -497,6 +475,17 @@ class ModmailBot(commands.Bot):
         )
         self.metadata_loop.before_loop(self.before_post_metadata)
         self.metadata_loop.start()
+
+        other_guilds = [
+            guild for guild in self.guilds if guild not in {self.guild, self.modmail_guild}
+        ]
+        if any(other_guilds):
+            logger.warning(
+                "The bot is in more servers other than the main and staff server."
+                "This may cause data compromise (%s).",
+                ", ".join(guild.name for guild in other_guilds),
+            )
+            logger.warning("If the external servers are valid, you may ignore this message.")
 
     async def convert_emoji(self, name: str) -> str:
         ctx = SimpleNamespace(bot=self, guild=self.modmail_guild)
@@ -935,7 +924,7 @@ class ModmailBot(commands.Bot):
                     return
                 await thread.recipient.trigger_typing()
 
-    async def handle_reaction_events(self, payload, *, add):
+    async def handle_reaction_events(self, payload):
         user = self.get_user(payload.user_id)
         if user.bot:
             return
@@ -961,7 +950,7 @@ class ModmailBot(commands.Bot):
             if not thread:
                 return
             if (
-                add
+                payload.event_type == "REACTION_ADD"
                 and message.embeds
                 and str(reaction) == str(close_emoji)
                 and self.config.get("recipient_thread_close")
@@ -992,7 +981,7 @@ class ModmailBot(commands.Bot):
                 logger.warning("Failed to find linked message for reactions: %s", e)
                 return
 
-        if add:
+        if payload.event_type == "REACTION_ADD":
             if await self.add_reaction(linked_message, reaction):
                 await self.add_reaction(message, reaction)
         else:
@@ -1003,26 +992,13 @@ class ModmailBot(commands.Bot):
                 logger.warning("Failed to remove reaction: %s", e)
 
     async def on_raw_reaction_add(self, payload):
-        await self.handle_reaction_events(payload, add=True)
+        await self.handle_reaction_events(payload)
 
     async def on_raw_reaction_remove(self, payload):
-        await self.handle_reaction_events(payload, add=False)
+        await self.handle_reaction_events(payload)
 
     async def on_guild_channel_delete(self, channel):
         if channel.guild != self.modmail_guild:
-            return
-
-        try:
-            audit_logs = self.modmail_guild.audit_logs()
-            entry = await audit_logs.find(lambda a: a.target == channel)
-            mod = entry.user
-        except AttributeError as e:
-            # discord.py broken implementation with discord API
-            # TODO: waiting for dpy
-            logger.warning("Failed to retrieve audit log: %s.", e)
-            return
-
-        if mod == self.user:
             return
 
         if isinstance(channel, discord.CategoryChannel):
@@ -1039,6 +1015,19 @@ class ModmailBot(commands.Bot):
             logger.info("Log channel deleted.")
             self.config.remove("log_channel_id")
             await self.config.update()
+            return
+
+        audit_logs = self.modmail_guild.audit_logs(
+            limit=10, action=discord.AuditLogAction.channel_delete
+        )
+        entry = await audit_logs.find(lambda a: int(a.target.id) == channel.id)
+
+        if entry is None:
+            logger.debug("Cannot find the audit log entry for channel delete of %d.", channel.id)
+            return
+
+        mod = entry.user
+        if mod == self.user:
             return
 
         thread = await self.threads.find(channel=channel)
@@ -1068,8 +1057,13 @@ class ModmailBot(commands.Bot):
 
     async def on_message_delete(self, message):
         """Support for deleting linked messages"""
-        # TODO: use audit log to check if modmail deleted the message
+
+        if message.is_system():
+            return
+
         if isinstance(message.channel, discord.DMChannel):
+            if message.author == self.user:
+                return
             thread = await self.threads.find(recipient=message.author)
             if not thread:
                 return
@@ -1077,24 +1071,46 @@ class ModmailBot(commands.Bot):
                 message = await thread.find_linked_message_from_dm(message)
             except ValueError as e:
                 if str(e) != "Thread channel message not found.":
-                    logger.warning("Failed to find linked message to delete: %s", e)
+                    logger.debug("Failed to find linked message to delete: %s", e)
                 return
             embed = message.embeds[0]
             embed.set_footer(text=f"{embed.footer.text} (deleted)", icon_url=embed.footer.icon_url)
             await message.edit(embed=embed)
             return
 
+        if message.author != self.user:
+            return
+
         thread = await self.threads.find(channel=message.channel)
         if not thread:
             return
+
+        audit_logs = self.modmail_guild.audit_logs(
+            limit=10, action=discord.AuditLogAction.message_delete
+        )
+
+        entry = await audit_logs.find(lambda a: a.target == self.user)
+
+        if entry is None:
+            return
+
         try:
             await thread.delete_message(message, note=False)
+            embed = discord.Embed(
+                description="Successfully deleted message.", color=self.main_color
+            )
         except ValueError as e:
             if str(e) not in {"DM message not found.", "Malformed thread message."}:
-                logger.warning("Failed to find linked message to delete: %s", e)
-            return
+                logger.debug("Failed to find linked message to delete: %s", e)
+                embed = discord.Embed(
+                    description="Failed to delete message.", color=self.error_color
+                )
+            else:
+                return
         except discord.NotFound:
             return
+        embed.set_footer(text=f"Message ID: {message.id} from {message.author}.")
+        return await message.channel.send(embed=embed)
 
     async def on_bulk_message_delete(self, messages):
         await discord.utils.async_all(self.on_message_delete(msg) for msg in messages)
@@ -1161,45 +1177,17 @@ class ModmailBot(commands.Bot):
                             corrected_permission_level.name,
                         )
             logger.warning("CheckFailure: %s", exception)
+        elif isinstance(exception, commands.DisabledCommand):
+            logger.info(
+                "DisabledCommand: %s is trying to run eval but it's disabled", context.author.name
+            )
         else:
             logger.error("Unexpected exception:", exc_info=exception)
 
-    async def validate_database_connection(self):
-        try:
-            await self.db.command("buildinfo")
-        except Exception as exc:
-            logger.critical("Something went wrong while connecting to the database.")
-            message = f"{type(exc).__name__}: {str(exc)}"
-            logger.critical(message)
-
-            if "ServerSelectionTimeoutError" in message:
-                logger.critical(
-                    "This may have been caused by not whitelisting "
-                    "IPs correctly. Make sure to whitelist all "
-                    "IPs (0.0.0.0/0) https://i.imgur.com/mILuQ5U.png"
-                )
-
-            if "OperationFailure" in message:
-                logger.critical(
-                    "This is due to having invalid credentials in your MONGO_URI. "
-                    "Remember you need to substitute `<password>` with your actual password."
-                )
-                logger.critical(
-                    "Be sure to URL encode your username and password (not the entire URL!!), "
-                    "https://www.urlencoder.io/, if this issue persists, try changing your username and password "
-                    "to only include alphanumeric characters, no symbols."
-                    ""
-                )
-            raise
-        else:
-            logger.debug("Successfully connected to the database.")
-        logger.line("debug")
-
     async def post_metadata(self):
-        owner = (await self.application_info()).owner
+        info = await self.application_info()
+
         data = {
-            "owner_name": str(owner),
-            "owner_id": owner.id,
             "bot_id": self.user.id,
             "bot_name": str(self.user),
             "avatar_url": str(self.user.avatar_url),
@@ -1212,6 +1200,19 @@ class ModmailBot(commands.Bot):
             "selfhosted": True,
             "last_updated": str(datetime.utcnow()),
         }
+
+        if info.team is not None:
+            data.update(
+                {
+                    "owner_name": info.team.owner.name
+                    if info.team.owner is not None
+                    else "No Owner",
+                    "owner_id": info.team.owner_id,
+                    "team": True,
+                }
+            )
+        else:
+            data.update({"owner_name": info.owner.name, "owner_id": info.owner.id, "team": False})
 
         async with self.session.post("https://api.logviewer.tech/metadata", json=data):
             logger.debug("Uploading metadata to Modmail server.")
